Rewrites tests and reveals a flaw
In order to fulfill the goal of full test coverage, this commit rewrites
and adds tests for both the clock and time_receiver.

The tests revealed a flaw in how tokio's bounded channels work here.
Adding more than a channel can hold causes the next receiver to return
an error. Once a receiver gets the error it clears all overflowed items
from the channel and leaves an amount of items equal to the channel's
size. The items left are the latest ones put in the channel.