Implements 'safe_time' and adds error handling
In order to make the clock safer, this commit adds the safe_time method
that will return any error that may occur when getting the time. Error
propagation was also added to all the methods that didn't already have
it, except time().

In addition to the error handling, this commit also adds a check for if
the clock is running or not before getting the time.